[DAT-6629] Refactoring rollbackToDAte command.
