+++ liquibase-core/src/main/java/liquibase/command/core/RollbackToDateCommandStep.java
@@ -1,73 +1,51 @@
 package liquibase.command.core;
 
+import liquibase.Scope;
+import liquibase.changelog.RanChangeSet;
+import liquibase.changelog.filter.ExecutedAfterChangeSetFilter;
 import liquibase.command.*;
-import liquibase.configuration.ConfigurationValueObfuscator;
-import liquibase.exception.CommandExecutionException;
+import liquibase.database.Database;
+import liquibase.logging.mdc.MdcKey;
 
-import java.time.LocalDateTime;
+import java.util.Date;
+import java.util.List;
 
-public class RollbackToDateCommandStep extends AbstractCliWrapperCommandStep {
+public class RollbackToDateCommandStep extends AbstractRollbackCommandStep {
 
     public static final String[] COMMAND_NAME = {"rollbackToDate"};
 
-    public static final CommandArgumentDefinition<String> CHANGELOG_FILE_ARG;
-    public static final CommandArgumentDefinition<String> URL_ARG;
-    public static final CommandArgumentDefinition<String> DEFAULT_SCHEMA_NAME_ARG;
-    public static final CommandArgumentDefinition<String> DEFAULT_CATALOG_NAME_ARG;
-    public static final CommandArgumentDefinition<String> USERNAME_ARG;
-    public static final CommandArgumentDefinition<String> PASSWORD_ARG;
-    public static final CommandArgumentDefinition<String> LABEL_FILTER_ARG;
-    public static final CommandArgumentDefinition<String> CONTEXTS_ARG;
-    public static final CommandArgumentDefinition<String> ROLLBACK_SCRIPT_ARG;
-    public static final CommandArgumentDefinition<LocalDateTime> DATE_ARG;
-    public static final CommandArgumentDefinition<String> CHANGE_EXEC_LISTENER_CLASS_ARG;
-    public static final CommandArgumentDefinition<String> CHANGE_EXEC_LISTENER_PROPERTIES_FILE_ARG;
-    public static final CommandArgumentDefinition<String> DRIVER_ARG;
-    public static final CommandArgumentDefinition<String> DRIVER_PROPERTIES_FILE_ARG;
+    public static final CommandArgumentDefinition<Date> DATE_ARG;
 
     static {
         CommandBuilder builder = new CommandBuilder(COMMAND_NAME);
-        URL_ARG = builder.argument(CommonArgumentNames.URL, String.class).required()
-                .description("The JDBC database connection URL").build();
-        DEFAULT_SCHEMA_NAME_ARG = builder.argument("defaultSchemaName", String.class)
-                .description("The default schema name to use for the database connection").build();
-        DEFAULT_CATALOG_NAME_ARG = builder.argument("defaultCatalogName", String.class)
-                .description("The default catalog name to use for the database connection").build();
-        DRIVER_ARG = builder.argument("driver", String.class)
-                .description("The JDBC driver class").build();
-        DRIVER_PROPERTIES_FILE_ARG = builder.argument("driverPropertiesFile", String.class)
-                .description("The JDBC driver properties file").build();
-        USERNAME_ARG = builder.argument(CommonArgumentNames.USERNAME, String.class)
-                .description("Username to use to connect to the database").build();
-        PASSWORD_ARG = builder.argument(CommonArgumentNames.PASSWORD, String.class)
-                .description("Password to use to connect to the database")
-                .setValueObfuscator(ConfigurationValueObfuscator.STANDARD)
-                .build();
-        CHANGELOG_FILE_ARG = builder.argument(CommonArgumentNames.CHANGELOG_FILE, String.class).required()
-                .description("File to write changelog to").build();
-        LABEL_FILTER_ARG = builder.argument("labelFilter", String.class)
-                .addAlias("labels")
-                .description("Changeset labels to match").build();
-        CONTEXTS_ARG = builder.argument("contexts", String.class)
-                .description("Changeset contexts to match").build();
-        ROLLBACK_SCRIPT_ARG = builder.argument("rollbackScript", String.class)
-                .description("Rollback script to execute").build();
-        DATE_ARG = builder.argument("date", LocalDateTime.class).required()
+        DATE_ARG = builder.argument("date", Date.class).required()
             .description("Date to rollback changes to").build();
-        CHANGE_EXEC_LISTENER_CLASS_ARG = builder.argument("changeExecListenerClass", String.class)
-            .description("Fully-qualified class which specifies a ChangeExecListener").build();
-        CHANGE_EXEC_LISTENER_PROPERTIES_FILE_ARG = builder.argument("changeExecListenerPropertiesFile", String.class)
-            .description("Path to a properties file for the ChangeExecListenerClass").build();
+
+        builder.addArgument(AbstractRollbackCommandStep.ROLLBACK_SCRIPT_ARG).build();
+    }
+
+    @Override
+    public void run(CommandResultsBuilder resultsBuilder) throws Exception {
+        Scope.getCurrentScope().addMdcValue(MdcKey.LIQUIBASE_OPERATION, COMMAND_NAME[0]);
+        Scope.getCurrentScope().addMdcValue(MdcKey.LIQUIBASE_COMMAND_NAME, COMMAND_NAME[0]);
+        CommandScope commandScope = resultsBuilder.getCommandScope();
+        Date dateToRollBackTo = commandScope.getArgumentValue(DATE_ARG);
+        Scope.getCurrentScope().addMdcValue(MdcKey.ROLLBACK_TO_DATE, dateToRollBackTo.toString());
+
+        Database database = (Database) commandScope.getDependency(Database.class);
+
+        List<RanChangeSet> ranChangeSetList = database.getRanChangeSetList();
+        this.doRollback(resultsBuilder, new ExecutedAfterChangeSetFilter(dateToRollBackTo, ranChangeSetList), ranChangeSetList);
+    }
+
+    @Override
+    protected String getOperationCommand() {
+        return "rollbackToDate";
     }
 
     @Override
     public String[][] defineCommandNames() {
         return new String[][] { COMMAND_NAME };
-    }
-
-    @Override
-    protected String[] collectArguments(CommandScope commandScope) throws CommandExecutionException {
-        return collectArguments(commandScope, null, "date");
     }
 
     @Override
